[Fix] Fix VLADBench
